+++ src/jobs/ImportJob.php
@@ -11,33 +11,16 @@ use DateTime;
 use putyourlightson\campaign\Campaign;
 use putyourlightson\campaign\events\ImportEvent;
 use putyourlightson\campaign\services\ImportsService;
-use yii\queue\RetryableJobInterface;
 
 /**
  * @property-read int $ttr
  */
-class ImportJob extends BaseJob implements RetryableJobInterface
+class ImportJob extends BaseJob
 {
     /**
      * @var int
      */
     public int $importId;
-
-    /**
-     * @inheritdoc
-     */
-    public function getTtr(): int
-    {
-        return Campaign::$plugin->settings->importJobTtr;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function canRetry($attempt, $error): bool
-    {
-        return false;
-    }
 
     /**
      * @inheritdoc