+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## 2.11.1 - Unreleased
 
+### Added
+
+- Added a `syncJobPriority` config setting that determines what priority to give sync jobs.
+
 ### Fixed
 
 - Fixed a bug in which the element count was not appearing for campaigns and mailing lists when there were no results.
 - Fixed a bug that could cause an error when viewing contacts in an incomplete segment ([#441](https://github.com/putyourlightson/craft-campaign/issues/441)).
+- Fixed a bug in which the import job TTR setting was not being respected ([#445](https://github.com/putyourlightson/craft-campaign/issues/445)).
+- Fixed a bug in which the sendout job max retry attempts setting was not being respected.
 
 ## 2.11.0 - 2023-11-28